youtube_sentiment/requirements.txt
python_version>=3.7
django==3.2.19
google-api-python-client==2.84.0
nltk==3.8.1
numpy==1.19.5
plotly==5.13.1
wordcloud==1.9.2
matplotlib==3.1.1
django-environ==0.10.0
djangorestframework==3.14.0
coreapi==2.3.3
pillow==9.5.0
